+++ libs/openFrameworks/graphics/ofPath.cpp
@@ -60,9 +60,19 @@ ofPath::ofPath(){
 }
 
 //----------------------------------------------------------
-ofPath::ofPath(const ofPolyline &polyline) : ofPath::ofPath() {
-	mode = POLYLINES;
-	polylines.push_back(polyline);
+ofPath::ofPath(const ofPolyline &polyline) : ofPath(){
+	addPolyline(polyline);
+}
+
+//----------------------------------------------------------
+void ofPath::addPolyline(const ofPolyline &polyline){
+	if (mode == COMMANDS) {
+		for (ofVec3f vertex : polyline) {
+			lineTo(vertex.x, vertex.y, vertex.z);
+		}
+	} else if (mode == POLYLINES) {
+		polylines.push_back(polyline);
+	}
 }
 
 //----------------------------------------------------------