+++ CHANGELOG.md
@@ -21,7 +21,8 @@
  / modified
 
 ### graphics
-    + ofPath has a polyline constructor
+    + ofPath has a polyline constructor and an addPolyline method for faster
+      ofPolyline -> ofPath conversions
 
  ------------------------------------------------------------------------------